Extract roleplay encounter UI components
