+++ apps/web/app/(app)/encounters/[id]/RoleplayEncounterScreens.test.ts
@@ -4,14 +4,33 @@ import { fileURLToPath } from "node:url";
 import { describe, expect, it } from "vitest";
 
 const sourcePath = join(dirname(fileURLToPath(import.meta.url)), "RoleplayEncounterScreens.tsx");
+const componentsPath = join(dirname(fileURLToPath(import.meta.url)), "components");
+const componentSourcePaths = [
+  join(componentsPath, "RoleplayCalculationPanel.tsx"),
+  join(componentsPath, "RoleplayRollBlock.tsx"),
+  join(componentsPath, "RoleplaySections.tsx"),
+  join(componentsPath, "roleplayRollTypes.ts"),
+  join(componentsPath, "roleplayStyles.ts"),
+];
 
 function readSource(): string {
   return readFileSync(sourcePath, "utf8");
 }
 
+function readComponentSource(fileName: string): string {
+  return readFileSync(join(componentsPath, fileName), "utf8");
+}
+
+function readRoleplaySources(): string {
+  return [
+    readSource(),
+    ...componentSourcePaths.map((path) => readFileSync(path, "utf8")),
+  ].join("\n");
+}
+
 describe("RoleplayEncounterScreens", () => {
   it("renders GM top info, GM message, visibility, roster descriptions, skill rolls, and action log", () => {
-    const source = readSource();
+    const source = readRoleplaySources();
 
     expect(source).toContain("Type: Roleplaying");
     expect(source).toContain("GM message");
@@ -44,7 +63,7 @@ describe("RoleplayEncounterScreens", () => {
   });
 
   it("stores roleplay state through sessionJson helpers and evaluates non-opposed success levels", () => {
-    const source = readSource();
+    const source = readRoleplaySources();
 
     expect(source).toContain("normalizeRoleplayState");
     expect(source).toContain("updateRoleplayGmMessage");
@@ -63,10 +82,15 @@ describe("RoleplayEncounterScreens", () => {
 
   it("keeps player roleplay screen minimal and hides GM-only sections", () => {
     const source = readSource();
-    const playerSource = source.slice(source.indexOf("export function PlayerRoleplayingEncounterScreen"));
+    const sectionsSource = readComponentSource("RoleplaySections.tsx");
+    const playerSource = [
+      source.slice(source.indexOf("export function PlayerRoleplayingEncounterScreen")),
+      sectionsSource.slice(sectionsSource.indexOf("export function PlayerRoleplayEncounterView")),
+    ].join("\n");
 
     expect(playerSource).toContain("Roleplaying encounter player tools will appear here.");
-    expect(playerSource).toContain("<RoleplayTopInfo");
+    expect(playerSource).toContain("<PlayerRoleplayEncounterView");
+    expect(playerSource).toContain("<EncounterInfoCard");
     expect(playerSource).not.toContain("Visibility grid");
     expect(playerSource).not.toContain("Roleplay roster descriptions");
     expect(playerSource).not.toContain("Skill roll assignment");
@@ -74,7 +98,7 @@ describe("RoleplayEncounterScreens", () => {
   });
 
   it("orders and labels roleplay participants and difficulty options", () => {
-    const source = readSource();
+    const source = readRoleplaySources();
 
     expect(source).toContain("orderRoleplayEncounterParticipants");
     expect(source).toContain("roleplayDifficultyOptions");
@@ -85,7 +109,7 @@ describe("RoleplayEncounterScreens", () => {
   });
 
   it("renders opposed roll state, support metadata, and opposed results in GM-only areas", () => {
-    const source = readSource();
+    const source = readRoleplaySources();
 
     expect(source).toContain("mode: \"opposed\"");
     expect(source).toContain("mode: \"difficulty\"");
@@ -101,15 +125,15 @@ describe("RoleplayEncounterScreens", () => {
     expect(source).toContain("opponentBlockOpen: nextDifficulty === \"none\" ? draft.opponentBlockOpen : false");
     expect(source).toContain("opponentBlockOpen: false");
     expect(source).toContain("opponentBlockOpen: true");
-    expect(source).toContain("handleAssignSkillRoll(draft, \"actor\")");
-    expect(source).toContain("handleAssignSkillRoll(draft, \"opponent\")");
-    expect(source).toContain("handleGmRoll(draft, \"actor\")");
-    expect(source).toContain("handleGmRoll(draft, \"opponent\")");
-    expect(source).toContain("handleGmRoll(draft, \"both\")");
+    expect(source).toContain("onAssignSkillRoll(draft, \"actor\")");
+    expect(source).toContain("onAssignSkillRoll(draft, \"opponent\")");
+    expect(source).toContain("onGmRoll(draft, \"actor\")");
+    expect(source).toContain("onGmRoll(draft, \"opponent\")");
+    expect(source).toContain("onGmRoll(draft, \"both\")");
   });
 
   it("keeps the roll preview concise with pending notes outside the formula", () => {
-    const source = readSource();
+    const source = readRoleplaySources();
 
     expect(source).toContain("pendingModifierLabels");
     expect(source).toContain("formulaText");
@@ -123,10 +147,11 @@ describe("RoleplayEncounterScreens", () => {
   });
 
   it("shows ranked rows only through total and keeps fumble details out of ranked rows", () => {
-    const source = readSource();
-    const rankedSource = source.slice(
-      source.indexOf("<h2 style={{ margin: 0 }}>Ranked roll results</h2>"),
-      source.indexOf("<h2 style={{ margin: 0 }}>Action log</h2>")
+    const source = readRoleplaySources();
+    const sectionsSource = readComponentSource("RoleplaySections.tsx");
+    const rankedSource = sectionsSource.slice(
+      sectionsSource.indexOf("<h2 style={{ margin: 0 }}>Ranked roll results</h2>"),
+      sectionsSource.indexOf("<h2 style={{ margin: 0 }}>Action log</h2>")
     );
 
     expect(rankedSource).toContain("entry.numericSubtotal == null ? \"unresolved\" : `total ${entry.numericSubtotal}`");
@@ -138,7 +163,7 @@ describe("RoleplayEncounterScreens", () => {
   });
 
   it("locks actor and opponent roll controls independently after each side rolls", () => {
-    const source = readSource();
+    const source = readRoleplaySources();
 
     expect(source).toContain("const actorLocked = Boolean(draft.actorRoll)");
     expect(source).toContain("const opponentLocked = Boolean(draft.opponentRoll)");
@@ -150,7 +175,7 @@ describe("RoleplayEncounterScreens", () => {
   });
 
   it("uses persistent two-column roll blocks with structured calculation panels", () => {
-    const source = readSource();
+    const source = readRoleplaySources();
 
     expect(source).toContain("gridTemplateColumns: \"minmax(0, 1fr) minmax(22rem, 1fr)\"");
     expect(source).toContain("alignSelf: \"stretch\"");
@@ -163,14 +188,14 @@ describe("RoleplayEncounterScreens", () => {
     expect(source).toContain("minHeight: \"10.5rem\"");
     expect(source).toContain("whiteSpace: \"nowrap\"");
     expect(source).toContain("<strong>{label}:</strong> {preview?.formulaText ?? \"—\"}");
-    expect(source).toContain("function RoleplayRollCalculationPanel");
+    expect(source).toContain("function RoleplayCalculationPanel");
     expect(source).toContain("style={rollControlRowStyle}");
     expect(source).toContain("style={{ ...compactInputStyle, width: \"4.5rem\" }}");
     expect(source).not.toContain("<RoleplayRollPreviewPanel");
   });
 
   it("preselects the chosen opponent block participant without auto-selecting an opponent skill", () => {
-    const source = readSource();
+    const source = readRoleplaySources();
 
     expect(source).toContain("opponentParticipantId: event.target.value");
     expect(source).toContain("opponentRoll: undefined");
@@ -181,7 +206,7 @@ describe("RoleplayEncounterScreens", () => {
   });
 
   it("clears draft-only roll editor state without touching persisted logs", () => {
-    const source = readSource();
+    const source = readRoleplaySources();
 
     expect(source).toContain("function resetRollDrafts()");
     expect(source).toContain("setRollDrafts([");
@@ -195,7 +220,7 @@ describe("RoleplayEncounterScreens", () => {
   });
 
   it("uses all canonical ordinary and secondary skills instead of only participant-known skills", () => {
-    const source = readSource();
+    const source = readRoleplaySources();
 
     expect(source).toContain("input.content.skills");
     expect(source).toContain("skill.category === \"ordinary\" || skill.category === \"secondary\"");
